project-1: Use 32bit qout instead of 16bit

+++ project-1/cpu/include/fpga_api.h
@@ -18,7 +18,7 @@ private:
   unsigned int *output_;
   int8_t *qvec_;
   int8_t *qmat_;
-  short *qout_;
+  int32_t *qout_;
   int8_t *qdata_;
 
   int m_size_;
